+++ databento/historical/api/symbology.py
@@ -9,7 +9,7 @@ from requests import Response
 from databento.common.enums import Dataset
 from databento.common.parsing import datetime_to_date_string
 from databento.common.parsing import optional_date_to_string
-from databento.common.parsing import optional_symbols_list_to_string
+from databento.common.parsing import optional_symbols_list_to_list
 from databento.common.validation import validate_enum
 from databento.common.validation import validate_semantic_string
 from databento.historical.api import API_VERSION
@@ -65,10 +65,10 @@ class SymbologyHttpAPI(BentoHttpAPI):
 
         """
         stype_in_valid = validate_enum(stype_in, SType, "stype_in")
-        symbols_list = optional_symbols_list_to_string(symbols, stype_in_valid)
+        symbols_list = optional_symbols_list_to_list(symbols, stype_in_valid)
         data: dict[str, object | None] = {
             "dataset": validate_semantic_string(dataset, "dataset"),
-            "symbols": symbols_list,
+            "symbols": ",".join(symbols_list),
             "stype_in": str(stype_in_valid),
             "stype_out": str(validate_enum(stype_out, SType, "stype_out")),
             "start_date": datetime_to_date_string(start_date),