+++ tests/test_common_parsing.py
@@ -10,7 +10,7 @@ import pytest
 from databento.common.parsing import optional_date_to_string
 from databento.common.parsing import optional_datetime_to_string
 from databento.common.parsing import optional_datetime_to_unix_nanoseconds
-from databento.common.parsing import optional_symbols_list_to_string
+from databento.common.parsing import optional_symbols_list_to_list
 from databento.common.parsing import optional_values_list_to_string
 from databento_dbn import SType
 
@@ -47,26 +47,19 @@ def test_maybe_values_list_to_string_given_valid_inputs_returns_expected(
     # Assert
     assert result == expected
 
-
-def test_maybe_symbols_list_to_string_given_invalid_input_raises_type_error() -> None:
-    # Arrange, Act, Assert
-    with pytest.raises(TypeError):
-        optional_symbols_list_to_string(INCORRECT_TYPE, SType.RAW_SYMBOL)
-
-
 @pytest.mark.parametrize(
     "stype, symbols, expected",
     [
-        pytest.param(SType.RAW_SYMBOL, None, "ALL_SYMBOLS"),
-        pytest.param(SType.PARENT, "ES.fut", "ES.FUT"),
-        pytest.param(SType.PARENT, "ES,CL", "ES,CL"),
-        pytest.param(SType.PARENT, "ES,CL,", "ES,CL"),
-        pytest.param(SType.PARENT, "es,cl,", "ES,CL"),
-        pytest.param(SType.PARENT, ["ES", "CL"], "ES,CL"),
-        pytest.param(SType.PARENT, ["es", "cl"], "ES,CL"),
-        pytest.param(SType.CONTINUOUS, ["ES.N.0", "CL.n.0"], "ES.n.0,CL.n.0"),
-        pytest.param(SType.CONTINUOUS, ["ES.N.0", ["ES,cl"]], "ES.n.0,ES,CL"),
-        pytest.param(SType.CONTINUOUS, ["ES.N.0", "ES,cl"], "ES.n.0,ES,CL"),
+        pytest.param(SType.RAW_SYMBOL, None, ["ALL_SYMBOLS"]),
+        pytest.param(SType.PARENT, "ES.fut", ["ES.FUT"]),
+        pytest.param(SType.PARENT, "ES,CL", ["ES", "CL"]),
+        pytest.param(SType.PARENT, "ES,CL,", ["ES", "CL"]),
+        pytest.param(SType.PARENT, "es,cl,", ["ES", "CL"]),
+        pytest.param(SType.PARENT, ["ES", "CL"], ["ES", "CL"]),
+        pytest.param(SType.PARENT, ["es", "cl"], ["ES", "CL"]),
+        pytest.param(SType.CONTINUOUS, ["ES.N.0", "CL.n.0"], ["ES.n.0", "CL.n.0"]),
+        pytest.param(SType.CONTINUOUS, ["ES.N.0", ["ES,cl"]], ["ES.n.0", "ES", "CL"]),
+        pytest.param(SType.CONTINUOUS, ["ES.N.0", "ES,cl"], ["ES.n.0", "ES", "CL"]),
         pytest.param(SType.CONTINUOUS, "", ValueError),
         pytest.param(SType.CONTINUOUS, [""], ValueError),
         pytest.param(SType.CONTINUOUS, ["ES.N.0", ""], ValueError),
@@ -74,27 +67,35 @@ def test_maybe_symbols_list_to_string_given_invalid_input_raises_type_error() ->
         pytest.param(SType.PARENT, 123458, ValueError),
     ],
 )
-def test_optional_symbols_list_to_string_given_valid_inputs_returns_expected(
+def test_optional_symbols_list_to_list_given_valid_inputs_returns_expected(
     stype: SType,
     symbols: list[str] | None,
-    expected: str | type[Exception],
+    expected: list[object] | type[Exception],
 ) -> None:
     # Arrange, Act, Assert
-    if isinstance(expected, str):
-        assert optional_symbols_list_to_string(symbols, stype) == expected
+    if isinstance(expected, list):
+        assert optional_symbols_list_to_list(symbols, stype) == expected
     else:
         with pytest.raises(expected):
-            optional_symbols_list_to_string(symbols, stype)
+            optional_symbols_list_to_list(symbols, stype)
 
 
 @pytest.mark.parametrize(
     "symbols, stype, expected",
     [
-        pytest.param(12345, SType.INSTRUMENT_ID, "12345"),
-        pytest.param("67890", SType.INSTRUMENT_ID, "67890"),
-        pytest.param([12345, "  67890"], SType.INSTRUMENT_ID, "12345,67890"),
-        pytest.param([12345, [67890, 66]], SType.INSTRUMENT_ID, "12345,67890,66"),
-        pytest.param([12345, "67890,66"], SType.INSTRUMENT_ID, "12345,67890,66"),
+        pytest.param(12345, SType.INSTRUMENT_ID, ["12345"]),
+        pytest.param("67890", SType.INSTRUMENT_ID, ["67890"]),
+        pytest.param([12345, "  67890"], SType.INSTRUMENT_ID, ["12345", "67890"]),
+        pytest.param(
+            [12345, [67890, 66]],
+            SType.INSTRUMENT_ID,
+            ["12345", "67890", "66"],
+        ),
+        pytest.param(
+            [12345, "67890,66"],
+            SType.INSTRUMENT_ID,
+            ["12345", "67890", "66"],
+        ),
         pytest.param("", SType.INSTRUMENT_ID, ValueError),
         pytest.param([12345, ""], SType.INSTRUMENT_ID, ValueError),
         pytest.param([12345, [""]], SType.INSTRUMENT_ID, ValueError),
@@ -103,10 +104,10 @@ def test_optional_symbols_list_to_string_given_valid_inputs_returns_expected(
         pytest.param(12345, SType.CONTINUOUS, ValueError),
     ],
 )
-def test_optional_symbols_list_to_string_int(
+def test_optional_symbols_list_to_list_int(
     symbols: list[Number] | Number | None,
     stype: SType,
-    expected: str | type[Exception],
+    expected: list[object] | type[Exception],
 ) -> None:
     """
     Test that integers are allowed for SType.INSTRUMENT_ID.
@@ -114,44 +115,44 @@ def test_optional_symbols_list_to_string_int(
     If integers are given for a different SType we expect a ValueError.
 
     """
-    if isinstance(expected, str):
-        assert optional_symbols_list_to_string(symbols, stype) == expected
+    if isinstance(expected, list):
+        assert optional_symbols_list_to_list(symbols, stype) == expected
     else:
         with pytest.raises(expected):
-            optional_symbols_list_to_string(symbols, stype)
+            optional_symbols_list_to_list(symbols, stype)
 
 
 @pytest.mark.parametrize(
     "symbols, stype, expected",
     [
-        pytest.param(np.byte(120), SType.INSTRUMENT_ID, "120"),
-        pytest.param(np.short(32_000), SType.INSTRUMENT_ID, "32000"),
+        pytest.param(np.byte(120), SType.INSTRUMENT_ID, ["120"]),
+        pytest.param(np.short(32_000), SType.INSTRUMENT_ID, ["32000"]),
         pytest.param(
             [np.intc(12345), np.intc(67890)],
             SType.INSTRUMENT_ID,
-            "12345,67890",
+            ["12345", "67890"],
         ),
         pytest.param(
             [np.int_(12345), np.longlong(67890)],
             SType.INSTRUMENT_ID,
-            "12345,67890",
+            ["12345", "67890"],
         ),
         pytest.param(
             [np.int_(12345), np.longlong(67890)],
             SType.INSTRUMENT_ID,
-            "12345,67890",
+            ["12345", "67890"],
         ),
         pytest.param(
             [np.int_(12345), np.longlong(67890)],
             SType.INSTRUMENT_ID,
-            "12345,67890",
+            ["12345", "67890"],
         ),
     ],
 )
-def test_optional_symbols_list_to_string_numpy(
+def test_optional_symbols_list_to_list_numpy(
     symbols: list[Number] | Number | None,
     stype: SType,
-    expected: str | type[Exception],
+    expected: list[object] | type[Exception],
 ) -> None:
     """
     Test that weird numpy types are allowed for SType.INSTRUMENT_ID.
@@ -159,42 +160,46 @@ def test_optional_symbols_list_to_string_numpy(
     If integers are given for a different SType we expect a ValueError.
 
     """
-    if isinstance(expected, str):
-        assert optional_symbols_list_to_string(symbols, stype) == expected
+    if isinstance(expected, list):
+        assert optional_symbols_list_to_list(symbols, stype) == expected
     else:
         with pytest.raises(expected):
-            optional_symbols_list_to_string(symbols, stype)
+            optional_symbols_list_to_list(symbols, stype)
 
 
 @pytest.mark.parametrize(
     "symbols, stype, expected",
     [
-        pytest.param("NVDA", SType.RAW_SYMBOL, "NVDA"),
-        pytest.param(" nvda  ", SType.RAW_SYMBOL, "NVDA"),
-        pytest.param("NVDA,amd", SType.RAW_SYMBOL, "NVDA,AMD"),
-        pytest.param("NVDA,amd,NOC,", SType.RAW_SYMBOL, "NVDA,AMD,NOC"),
-        pytest.param("NVDA,  amd,NOC, ", SType.RAW_SYMBOL, "NVDA,AMD,NOC"),
-        pytest.param(["NVDA", ["NOC", "AMD"]], SType.RAW_SYMBOL, "NVDA,NOC,AMD"),
-        pytest.param(["NVDA", "NOC,AMD"], SType.RAW_SYMBOL, "NVDA,NOC,AMD"),
+        pytest.param("NVDA", SType.RAW_SYMBOL, ["NVDA"]),
+        pytest.param(" nvda  ", SType.RAW_SYMBOL, ["NVDA"]),
+        pytest.param("NVDA,amd", SType.RAW_SYMBOL, ["NVDA", "AMD"]),
+        pytest.param("NVDA,amd,NOC,", SType.RAW_SYMBOL, ["NVDA", "AMD", "NOC"]),
+        pytest.param("NVDA,  amd,NOC, ", SType.RAW_SYMBOL, ["NVDA", "AMD", "NOC"]),
+        pytest.param(
+            ["NVDA", ["NOC", "AMD"]],
+            SType.RAW_SYMBOL,
+            ["NVDA", "NOC", "AMD"],
+        ),
+        pytest.param(["NVDA", "NOC,AMD"], SType.RAW_SYMBOL, ["NVDA", "NOC", "AMD"]),
         pytest.param("", SType.RAW_SYMBOL, ValueError),
         pytest.param([""], SType.RAW_SYMBOL, ValueError),
         pytest.param(["NVDA", ""], SType.RAW_SYMBOL, ValueError),
         pytest.param(["NVDA", [""]], SType.RAW_SYMBOL, ValueError),
     ],
 )
-def test_optional_symbols_list_to_string_raw_symbol(
+def test_optional_symbols_list_to_list_raw_symbol(
     symbols: list[Number] | Number | None,
     stype: SType,
-    expected: str | type[Exception],
+    expected: list[object] | type[Exception],
 ) -> None:
     """
     Test that str are allowed for SType.RAW_SYMBOL.
     """
-    if isinstance(expected, str):
-        assert optional_symbols_list_to_string(symbols, stype) == expected
+    if isinstance(expected, list):
+        assert optional_symbols_list_to_list(symbols, stype) == expected
     else:
         with pytest.raises(expected):
-            optional_symbols_list_to_string(symbols, stype)
+            optional_symbols_list_to_list(symbols, stype)
 
 
 @pytest.mark.parametrize(